revno: 15829 message:   Fixed bug -  Programme validation rule exception with decimals.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Tue 2014-06-24 20:33:37 +0800

+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/program/DefaultProgramValidationService.java
@@ -158,8 +158,8 @@ public class DefaultProgramValidationService
                 String expression = validate.getLeftSide().getExpression() + " " + validate.getRightSide().getExpression();
                 if ( isNumberDataExpression( expression ) )
                 {
-                    int leftSide = Integer.parseInt( leftSideValue );
-                    int rightSide = Integer.parseInt( rightSideValue );
+                    double leftSide = Double.parseDouble( leftSideValue );
+                    double rightSide = Double.parseDouble( rightSideValue );
                     if ( !((operator.equals( "==" ) && leftSide == rightSide)
                         || (operator.equals( "<" ) && leftSide < rightSide)
                         || (operator.equals( "<=" ) && leftSide <= rightSide)
